Add support for custom ppa repopsitory and custom debian branch
- add new package definitions for focal and jammy
  rsyslog-mmdblookup
  rsyslog-debug
- fixed build of rsyslog-mmgrok package for jammy
  Added /usr/include/tirpc to CFLAGS so libgrok finds rpc/rpc.h
  Reason is: glibc >=2.32 dropped "--enable-obsolete-rpc"
- Fixed build for auto_daily scripts and added support for
  custom repositories.

closes: https://github.com/rsyslog/rsyslog-pkg-ubuntu/issues/163

+++ scripts/auto_daily.sh
@@ -1,28 +1,39 @@
-#!/bin/bash
+#! /bin/bash
 # Prequitsites: tarballs must exist in $INFRAHOME/repo/<project>
 # for all to be built projects. These tarballs must have been created
 # by "make dist". The tarball must be copied to this packages home
 # directory. Only one tarball must exist at the same time.
-echo build all daily packages
-date
-set -v
-set -x
+
+#set -o xtrace  # useful for debugging
+#set -v
+#set -x
 
 source $RSI_SCRIPTS/config.sh
 
-PPADAILY=daily-stable
-PPABRANCH=master
 PROJECTS=${1:-"libestr liblogging libfastjson liblognorm librelp rsyslog"}
-CUSTOMBUILD=${2}
+PPAREPO=$2	# daily-stable experimental
+PPABRANCH=$3	# v8-stable
+CUSTOMBUILD=$4	# Custom build if any
+
+# Abort the script if PPAREPO is not set
+if [ -z "$PPAREPO" ]; then
+    echo "Error: PPAREPO is not set."
+    exit 1
+fi
+
+echo BUILD Daily packages for $PROJECTS on REPO $PPAREPO with PPABRANCH $PPABRANCH
+date
 
 for PROJECT in $PROJECTS
 do
 	# Ubuntu (bionic 18.04, ) 20.04 and 22.04
 	for PLATFORM in focal jammy
 	do
+		echo BUILD $PROJECT for $PLATFORM
 		cd $INFRAHOME/repo/rsyslog-pkg-ubuntu/$PROJECT
 		rm -fv *.tar.gz
 		cp -v $INFRAHOME/repo/$PROJECT/*.tar.gz .
-		$INFRAHOME/repo/rsyslog-pkg-ubuntu/scripts/auto_daily_project.sh $PLATFORM $PPADAILY $PPABRANCH $CUSTOMBUILD
+		echo DIR $INFRAHOME/repo/rsyslog-pkg-ubuntu/$PROJECT RUN: $INFRAHOME/repo/rsyslog-pkg-ubuntu/scripts/auto_daily_project.sh $PLATFORM $PPAREPO $PPABRANCH $CUSTOMBUILD
+		$INFRAHOME/repo/rsyslog-pkg-ubuntu/scripts/auto_daily_project.sh $PLATFORM $PPAREPO $PPABRANCH $CUSTOMBUILD
 	done
 done
